requirements.txt
openai>=1.59.0
anthropic>=0.45.0
streamlit>=1.41.0
python-dotenv>=1.0.1
httpx>=0.27.0
requests>=2.31.0
exa-py>=1.0.0
trafilatura>=1.12.0
openpyxl>=3.1.5
gspread>=6.1.2
google-api-python-client>=2.170.0
google-auth>=2.40.0
